issue262: One more version of pagination
