+++ src/util.js
@@ -35,6 +35,7 @@ import { exportContext } from './util/exportContext.js'
 import { flatMap } from './util/flatMap.js'
 import { flatten } from './util/flatten.js'
 import { formatNumber } from './util/formatNumber.js'
+import { getChildPath } from './util/getChildPath.js'
 import { getComparisonToken } from './util/getComparisonToken.js'
 import { getContexts } from './util/getContexts.js'
 import { getContextsSortedAndRanked } from './util/getContextsSortedAndRanked.js'
@@ -155,6 +156,7 @@ export {
   flatMap,
   flatten,
   formatNumber,
+  getChildPath,
   getComparisonToken,
   getContexts,
   getContextsSortedAndRanked,
